# optimization: Accelerates the performance of updating the xDS configuration