@nlgwcy 2024-05-15
> Why `BenchmarkFlush`(actually listener) is quicker than cluster flush