@nlgwcy 2024-05-15
> Why `BenchmarkFlush`(actually listener) is quicker than cluster flush

from the prespective of the testcase design:
BenchmarkClusterFlush
Cluster records are continuously added to the map. (the keys of cluster map(cluster name) are different.)
```c
// cluster map define
struct {
    __uint(type, BPF_MAP_TYPE_HASH);
    __uint(key_size, CLUSTER_NAME_MAX_LEN);
    __uint(value_size, sizeof(Cluster__Cluster));
    __uint(map_flags, BPF_F_NO_PREALLOC);
    __uint(max_entries, MAP_SIZE_OF_CLUSTER);
} map_of_cluster SEC(".maps");
```
BenchmarkFlush:
Only one listener record is updated. (The `listener.Name` is randomly set, but the key of the listener map is `addr`.)
```c
// listener map define
struct {
    __uint(type, BPF_MAP_TYPE_HASH);
    __uint(key_size, sizeof(address_t));
    __uint(value_size, sizeof(Listener__Listener));
    __uint(max_entries, MAP_SIZE_OF_LISTENER);
    __uint(map_flags, BPF_F_NO_PREALLOC);
} map_of_listener SEC(".maps");
```