@hzxuzhonghu 2024-05-15
Good analyze, So we should update the listener  benchmark