@nlgwcy 2024-05-15
> Good analyze, So we should update the listener benchmark

may not be required. This case is used to test the scenario where the same record is updated. In `deserial_update_elem`, the instance tree corresponding to the record will be deleted first.